[rename] 파일명 수정, 폴더 구분 #84
